pyproject.toml
[tool.poetry]
name = "django-custom-multiple-filter"
version = "0.1.0"
description = ""
authors = ["moon <black2st11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
drf-yasg = "^1.21.3"
pytest-django = "^4.5.2"
pytest-mock = "^3.8.2"
requests-mock = "^1.9.3"
Faker = "^13.15.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
